# perf: Optimize storage memory by 85% - RASTAFARI OPTIMIZATION! 🔥