--- a/src-tauri/src/commands.rs
+++ b/src-tauri/src/commands.rs
@@ -366,15 +366,8 @@ pub async fn get_service_health_metrics(
 
             for service_id in service_ids {
                 if let Some(health) = storage.get_service_health(service_id) {
-                    // Resolve service name from string pool if available, else fallback
-                    let service_name = if let Some(ref string_pool) = state.string_pool {
-                        string_pool
-                            .get(service_id)
-                            .cloned()
-                            .unwrap_or_else(|| format!("service-{}", service_id))
-                    } else {
-                        format!("service-{}", service_id)
-                    };
+                    // Use service ID as placeholder (metric storage doesn't store names)
+                    let service_name = format!("service-{}", service_id);
                     result.push(ServiceHealth {
                         service_name,
                         request_rate: health.request_rate,